// src/app/@api/model/dealDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemIdDto } from './itemIdDto';
import { SupplierIdDto } from './supplierIdDto';
import { UserIdDto } from './userIdDto';


export interface DealDto { 
    creationDate?: Date;
    customer?: UserIdDto;
    dealId?: number;
    dealNumber?: number;
    item?: ItemIdDto;
    status?: DealDto.StatusEnum;
    supplier?: SupplierIdDto;
}
export namespace DealDto {
    export type StatusEnum = 'ACCEPTED' | 'CANCELED' | 'DRAFT' | 'INQUIRY_DONE' | 'INQUIRY_INPROGRESS' | 'INVOICE_INPROGRESS' | 'ORDER_DONE' | 'ORDER_INPROGRESS' | 'QUOTATION_DONE' | 'QUOTATION_INPROGRESS';
    export const StatusEnum = {
        ACCEPTED: 'ACCEPTED' as StatusEnum,
        CANCELED: 'CANCELED' as StatusEnum,
        DRAFT: 'DRAFT' as StatusEnum,
        INQUIRYDONE: 'INQUIRY_DONE' as StatusEnum,
        INQUIRYINPROGRESS: 'INQUIRY_INPROGRESS' as StatusEnum,
        INVOICEINPROGRESS: 'INVOICE_INPROGRESS' as StatusEnum,
        ORDERDONE: 'ORDER_DONE' as StatusEnum,
        ORDERINPROGRESS: 'ORDER_INPROGRESS' as StatusEnum,
        QUOTATIONDONE: 'QUOTATION_DONE' as StatusEnum,
        QUOTATIONINPROGRESS: 'QUOTATION_INPROGRESS' as StatusEnum
    };
}